tex mapping basics
- I have a tex mapped sphere but that is it
- i understand it now

+++ A5/include/NormalMaterial.hpp
@@ -10,7 +10,7 @@ class NormalMaterial : public Material
 
     virtual glm::dvec3 getRadiance(
         const Ray & ray, const Intersection & intersect,
-        const glm::vec3 & ambient, const std::vector<const Light *> & lights,
+        const glm::dvec3 & ambient, const std::vector<const Light *> & lights,
         const glm::dvec3 & reflectionDir = glm::dvec3(0.0),
         const glm::dvec3 & reflectionRadiance = glm::dvec3(0.0),
         const glm::dvec3 & refractionDir = glm::dvec3(0.0),
