+++ A5/libs/PhongMaterial.cpp
@@ -31,7 +31,7 @@ double PhongMaterial::getShininess() const
 }
 
 glm::dvec3 PhongMaterial::getRadiance(
-    const Ray & ray, const Intersection & intersect, const glm::vec3 & ambient,
+    const Ray & ray, const Intersection & intersect, const glm::dvec3 & ambient,
     const std::vector<const Light *> & lights, const glm::dvec3 & reflectionDir,
     const glm::dvec3 & reflectionRadiance, const glm::dvec3 & refractionDir,
     const glm::dvec3 & refractionRadiance) const
